+++ prisma/migrations/20210918183212_removed_default_value_of_data_source/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "MarketData" ALTER COLUMN "dataSource" DROP DEFAULT;
+
+-- AlterTable
+ALTER TABLE "Order" ALTER COLUMN "dataSource" DROP DEFAULT;