+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Removed the default value of the data source attribute
+
+### Todo
+
+- Apply data migration (`yarn prisma migrate deploy`)
+
 ## 1.54.0 - 18.09.2021
 
 ### Added
@@ -24,7 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Todo
 
-- Apply data migration (`yarn database:push`)
+- Apply data migration (`yarn prisma migrate deploy`)
 
 ## 1.53.0 - 13.09.2021
 
@@ -146,7 +156,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Todo
 
-- Apply data migration (`yarn database:push`)
+- Apply data migration (`yarn prisma migrate deploy`)
 
 ## 1.41.0 - 21.08.2021
 
@@ -199,7 +209,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Todo
 
-- Apply data migration (`yarn database:push`)
+- Apply data migration (`yarn prisma migrate deploy`)
 
 ## 1.38.0 - 14.08.2021
 
@@ -259,7 +269,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Todo
 
-- Apply data migration (`yarn database:push`)
+- Apply data migration (`yarn prisma migrate deploy`)
 
 ## 1.34.0 - 07.08.2021